oled-tools: remove release string from oled version
Make oled binary print major.minor as version. This is to avoid the
mismatch with spec file. The spec file would update the release string
often to include new bug fixes. With this patch, we do not need to
update the oled.py each time we update the spec file with new release
string.

Orabug: 31793813

Signed-off-by: Manjunath Patil <manjunath.b.patil@oracle.com>
Reviewed-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>