Smtool: Fix scan/verbose scan and unsupported variants list on Oracle VM Servers.
On one of the OVS machines, we found that scan function
recommended upgrades for variants for which mitigations are
not available.
Fix scan function so that appropriate recommendations
are issued.

Fix verbose scan to avoid redundant display of
cmdline options. Add iTLB Multihit to unsupported
variants list while enabling/disabling mitigations.

Orabug: 33005737
Signed-off-by: Mridula Shastry <mridula.c.shastry@oracle.com>
Reviewed-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>