Smtool: Avoid failure when mitigation for a particular variant is not available
Smtool operations like scan, enabling/disabling mitigation fails when
mitigation for a particular variant (validated by the presence of status
sysfile which is dependent on the kernel version) is not present
on a system:

[root@ol6-uek4 oled-tools]# oled smtool --enable-default-mitigations -y
Traceback (most recent call last):
  File "/usr/lib/oled-tools/smtool", line 488, in <module>
    SMT = Smtool(sys.argv)
  File "/usr/lib/oled-tools/smtool", line 465, in __init__
    self.host.scan_host()
  File "/usr/lib/python2.6/site-packages/smtool_lib/host.py", line 189, in
scan_host
    self.vuln.scan_vulnerabilities()
  File "/usr/lib/python2.6/site-packages/smtool_lib/vulnerabilities.py", line
760, in scan_vulnerabilities
    is_mitigation_possible = self.v_7.scan_variant()
  File "/usr/lib/python2.6/site-packages/smtool_lib/variant.py", line 649, in
scan_variant
    if os.path.exists(self.sysfile.get_sysfile()):
  File "/usr/lib64/python2.6/genericpath.py", line 18, in exists
    st = os.stat(path)
TypeError: coercing to Unicode: need string or buffer, NoneType found

The issue was because of the issue introduced by commit
23032d5f1089 ('Smtool: Add support for detection/mitigation of SRBDS')
which modified return statements in get_sysfile function. Fix get_sysfile
function and also add appropriate checks in functions in other modules
to to avoid the above stacktrace.

Orabug: 33287128
Signed-off-by: Mridula Shastry <mridula.c.shastry@oracle.com>
Reviewed-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>