oled-tools: Restructure tools directories
This commit creates a new "tools" subdir in oled-tools root, and
moves kcore-utils, kstack, lkce, and memstate subdirs under tools.
The file structures in the kcore-utils/kstack/lkce/memstate subdirs
themselves are unchanged. These are mere renames - no functional
changes.

The only code change in this commit is to the Makefile, so that
the build process is updated with the new directory structure.

Signed-off-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Signed-off-by: Manjunath Patil <manjunath.b.patil@oracle.com>