memstate: Add a margin of error for the MFK check/warning.
In the memstate default/summary output, it prints a warning if the
tunable vm.min_free_kbytes is not set to the recommended value
(note that this check might be removed in future versions of this
tool). This commit adds some leeway for that check, since
we don't care if the actual value deviates from the recommended
value by a small margin.

Signed-off-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Reviewed-by: Jose Lombera <jose.lombera@oracle.com>