memstate: Add a sleep while iterating through /proc/<pid>/smaps_rollup files
The '-p' option reports per-process memory usage by all processes
running on the system. To do that, it reads the smaps_rollup file for
each process, under /proc. This can be an expensive operation.
Jianfeng's tests found that on a system with 9764 process, `oled
memstate -p` consumed high sys CPU (> 98%) for over 66 seconds. This is
undesirable, esp. on a production system. To mitigate this, periodically
give up the CPU, say every 100 files. This will extend the runtime of
memstate, but the CPU sys utilization will come down.

Orabug: 36432089

Signed-off-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Reported-by: Jianfeng Wang <jianfeng.w.wang@oracle.com>
Reviewed-by: Jianfeng Wang <jianfeng.w.wang@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>