memstate: Reduce sleep duration and frequency for some memstate options
A few flows in memstate data processing sleep for 10 ms for every
/proc/<pid>/<file> read, in order to not hog the CPU and drive up %sys.
But that increases the execution time a lot, esp. on busy servers with
more than 10K processes. For instance, on a system with ~10K processes,
running `oled memstate -n -v` took about 912 seconds, which is 15
minutes. While this is not harmful to the system or other processes
running on the server, it can be confusing for the user who might wonder
why the tool's taking so long, or if it is stuck. Reduce this execution
time by sleeping 1 ms for every 100 procfs files processed.

Replacing string concatention with str.join() also speeds up the loops
significantly.

Orabug: 36432139

Signed-off-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Reported-by: Jianfeng Wang <jianfeng.w.wang@oracle.com>
Reviewed-by: Jianfeng Wang <jianfeng.w.wang@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>