memstate: Replace round() with formatted string for printing
At these places, memstate records the amount of time spent in the for
loop of parsing system information. The execution time is rounded in the
debug output. For example, in memstate_lib/rss.py, __display_top_vmrss()
parses all /proc/*/status files. On the testing server, the execution
time is either 0 or 1 in the memstate log file.

The fix is to replace round() with Python's formatted string for
printing the execution time with two digits after the decimal point.

Orabug: 36432149

Signed-off-by: Jianfeng Wang <jianfeng.w.wang@oracle.com>
Reviewed-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>