memstate: Improve __create_matrix() for hugepage processing
In memstate_lib/hugepages.py, __create_matrix() is used to turn the raw
data parsed from /sys/devices/system/node/node*/hugepages/hugepages-*
into a dictionary that contains KBs of hugepages (of each hugepage size)
for each NUMA node.

In the current version, __create_matrix() has an unused for-loop and a
comment section that does not match the code. This commit removes the
unused code section and also updates the comment section.

Orabug: 36432149

Signed-off-by: Jianfeng Wang <jianfeng.w.wang@oracle.com>
Reviewed-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>