oled-tools/syswatch: Invert the sense of the idle% threshold for -s
The -s flag to syswatch allows users to specify thresholds for CPU stats
such that if the CPU stat becomes greater than or equal to the threshold
value specified, a command the user gives will fire. This patch inverts
the sense of the threshold for the CPU idle percentage stat. It also
separates the CPU stats into two lists: 1) stats that will fire when
the CPU stat value is less than or equal to the threshold value provided,
and 2) stats that will fire when the CPU stat value is greater than or
equal to the threshold value provided. Currently idle is the only CPU stat
in the former list.

Previously 'oled syswatch -s idle:5 -c ls' would fire the ls command
if CPU the idle percentage became greater than or equal to 5%. This
patch changes the behavior such that 'oled syswatch -s idle:5 -c ls'
fires the ls command if the CPU idle percentage is less than or equal
to 5%.

Orabug: 36622809

Signed-off-by: Ryan McCabe <ryan.m.mccabe@oracle.com>
Reviewed-by Jeffery Yoder <jeffery.yoder@oracle.com>