oled/lkce: chroot before using `/usr/bin/timeout`
chroot to the LKCE environment before using /usr/bin/timeout.
This allows us to avoid having to add /usr/bin/timeout to the kexec initrd.
This also allows us to not have to write an 'extra_bins' config
parameter to /etc/kdump.conf, which simplifies operation and fixes a bug
that causes a duplicate line to sometimes be written, which can cause a
hang on crashes.

Resolves: LSE-386 and LSE-400

Signed-off-by: Ryan McCabe <ryan.m.mccabe@oracle.com>
Reviewed-by: Stephen Brennan <stephen.s.brennan@oracle.com>