oled-tools/lkce: Mount needed devices in kdump mode
In the kdump environment, LKCE needs both the sysroot and the LKCE report
output directory to be available in order to run crash or corelens against
/proc/vmcore and save the output.

This may entail mounting multiple devices if the LKCE dump directory
is not on the same filesystem as /. This patch will mount all needed
filesystems, either by a bind mount, if the device has already been
mounted by kdump, or by UUID, if it has not. This patch also activates
any underlying LVM volume groups needed to mount the necessary
directories.

This resolves LSE-382 and LSE-383

Signed-off-by: Ryan McCabe <ryan.m.mccabe@oracle.com>
Reviewed-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Jeffery Yoder <jeffery.yoder@oracle.com>