oled-tools/lkce: Try to insert fs modules if mount fails
If a mount fails because of unknown filesystem type, attempt to chroot
to /sysroot and modprobe the fs module. This fixes a corner case
where LKCE would fail if its output directory was on a filesystem
for which there is no module in the kdump initrd (e.g., ext4). This
approach allows us to not have to modify /etc/kdump.conf to add in
extra modules.

Resolves LSE-413

Signed-off-by: Ryan McCabe <ryan.m.mccabe@oracle.com>
Reviewed-by: Jeffery Yoder <jeffery.yoder@oracle.com>