oled-tools/lkce: Avoid generating files that are static
LKCE autogenerates /etc/oled/lkce/corelens_args_file and
/etc/oled/lkce/crash_cmds_file which are entirely static.
These files are now installed by the installer instead of
generated on the fly. LKCE also autogenerates the kdump_pre.sh
file. While we can't avoid generating some of this file, we can
move the bulk of it into a file that gets installed, and appended
to the parts that must be generated on the fly. This eases both
maintenance and readability, as we are no longer generating
the while shell script.

Resolves LSE-408

Signed-off-by: Ryan McCabe <ryan.m.mccabe@oracle.com>
Reviewed-by: Stephen Brennan <stephen.s.brennan@oracle.com>