oled-tools/lkce: Disable use of pre.d on OL8 and OL9
Disable the use of placing symlinks in /etc/kdump/pre.d/ on OL8 and
OL9. Scripts run from /etc/kdump/pre.d/ will not short-circuit the
execution chain if they exit with a non-zero status, and as a result,
they cannot inhibit the dumping of vmcore. This patch requires LKCE
be configured via a kdump_pre line in /etc/kdump.conf on OL8 and OL9,
which matches the way it is configured on OL7.

Reviewed-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Ryan McCabe <ryan.m.mccabe@oracle.com>