Fix vmcore_sz failing on 6.3 kernels
vmcore_sz sub-command on 6.3 and newer kernels throws the error
"ValueError: 'TYPE' is not in list"

This is because the makedumpfile on these kernels throws the following
warnings and the tool fails to process them.
"The kernel version is not supported.
The makedumpfile operation may be incomplete."

Fixing the script to process these warnings.

Orabug: 37327741

Signed-off-by: Srikanth C S <srikanth.c.s@oracle.com>
Reviewed-by: Partha Satapathy <partha.satapathy@oracle.com>