Error out if kexec-tools is not installed or configured
If the file /etc/kdump.conf does not exist:
$ python3 vmcore_sz.py
Cannot get the default dump level.

Please check if kdump is configured.
The file /etc/kdump.conf does not exist

If the package kexec-tools is not installed:
$ python3 vmcore_sz.py -d  11
Cannot run the command makedumpfile.
Please check if the kexec-tools package is installed and configured.

Orabug: 36039350

Signed-off-by: Srikanth C S <srikanth.c.s@oracle.com>
Reviewed-by: Arumugam Kolappan <aru.kolappan@oracle.com>