oled-tools/olprof : add olprof to trace and profile a workload
olprof genartes profiling scripts dynamically for a list of functions
representing a workload and excutes the profiling script.
The profiled data contains outputs of  entry and return traces of each
function listed in workload list. If a function is not traceable,
olprof will skip it and continue to trace other functions.

The oled olprof command mandates use of the  option, which ponits
to workload function list. The default workloads available are:

kern_cpuhp
libvirt_cpuhp
qemu_cpuhp

The command oled olprof with a pid as an argument, will trace the
specified process, otherwise will trace the kernel.

The command oled olprof with -d option, excutes in debug mode.

Orabug: 37618519

Signed-off-by: Partha Sarathi Satapathy <partha.satapathy@oracle.com>
Reviewed-by: Srikanth C S <srikanth.c.s@oracle.com>
Reviewed-By: Jeffery Yoder <jeffery.yoder@oracle.com>