oled-tools/oomwatch: Add a script to verify kills
Add a script, /etc/oled/oomwatch/verify_kill.sh, on whose exit code
kills are contingent. The script is passed 3 parameters: pid,
process name, and RSS usage in bytes. If the script exits with
status 0, the process will be killed. If it exits with a non-zero
status, stdout will be logged, and the process will not be killed.

Orabug: 37723296

Reviewed-by: Jeffery Yoder <jeffery.yoder@oracle.com>
Signed-off-by: Ryan McCabe <ryan.m.mccabe@oracle.com>